+++ packages/apps/src/constants.ts
@@ -1,0 +1,13 @@
+// Copyright 2017-2019 @polkadot/ui-app authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+export enum SideBarTransition {
+	COLLAPSED = 'COLLAPSED',
+	EXPANDED = 'EXPANDED',
+	EXPANDED_AND_MAXIMISED = 'EXPANDED_AND_MAXIMISED',
+	MINIMISED_AND_EXPANDED = 'MINIMISED_AND_EXPANDED'
+}
+
+export const SIDEBAR_MENU_THRESHOLD = 768;
+export const SIDEBAR_TRANSITION_DURATION = 300;
