+++ packages/ui-app/src/constants.ts
@@ -6,13 +6,3 @@ export enum BitLengthOption {
   CHAIN_SPEC = 128,
   NORMAL_NUMBERS = 32
 }
-
-export enum SideBarTransition {
-	COLLAPSED = 'COLLAPSED',
-	EXPANDED = 'EXPANDED',
-	EXPANDED_AND_MAXIMISED = 'EXPANDED_AND_MAXIMISED',
-	MINIMISED_AND_EXPANDED = 'MINIMISED_AND_EXPANDED'
-}
-
-export const SIDEBAR_MENU_THRESHOLD = 768;
-export const SIDEBAR_TRANSITION_DURATION = 300;
